Redesign CW decoding
The CW decoder now uses an overlapping FFT method to find CW tones
on the available spectrum. The detection for now is based on the SNR so
the decoder tries to perform a noise floor estimation first. Then it
splits the available spectrum in multiple subchannels. Each subchannel
decodes the corresponding spectrum independently from the others.