Start the implementation of the TX API
The TX API is following a similar approach with the RX,
using endoder classes that extend the base encoder() class.

This generic API is used by the frame_encoder() block
to encode the frame. The encoding is performed in a message
fashion. Users then should use PDU to tagged Stream,
Tagged Stream Multiply Length Tag and Burst Shaper blocks
to properly contruct a burst for transmission.