Integrate a variation of the Constellation Modulator
The GNU Radio Constellation Modulator block uses
the polyphase arbitrary resampler and RRC filter
taps to perform pulse shaping.

This introduces an offset to the tag propagated throughout
the stream. To deal with this issue, a small variation of
this block has been integrated into the gr-satnogs codebase.
Our custom block declares the sample delay of the RRC filter
to the GNU Radio scheduler, so the tags can be propagated
properly towards other blocks, eg the burst shaper.