Implement the Mode 6 decoder for the AX100 modem
The decoder tries to find a valid frame start and
stop of an AX.25 frame. After this it discards
any kind of AX.25 overhead and tries to decode
the payload only, which is encoded with the
Reed Solomon code. The CRC check of the
AX.25 is not performed, allowing the RS to
correct possible bit errors.

Signed-off-by: Manolis Surligas <surligas@gmail.com>