AX.100 Mode 6: Make available the AX.25 header too
Until now, the Mode 6 decoder produced only the RS codeblock,
becasue both the RS and the CRC32-C ensuring the validity of
the frame. However, the AX.25 header holds vital information
that can be used to descriminate between nearby spacecrafts
using the same modem. Now the decoder also makes available
this headder, but users should take into consideration that
errors may exists on these data.

Many will argue that the AX.25 CRC could be used too, but
this defeats the purpose of using a FEC.

Signed-off-by: Manolis Surligas <surligas@gmail.com>