Add a BER calculator block
This commits adds a BER simulation block.
Because most of the gr-satnogs blocks are
operating in a per-frame basis, it it quite
difficult to create BER simulations.

However, based on the FER (frame error rate),
the BER can be estimated quite accurately.

Example BER simulations can be found at the examples
directory.

Signed-off-by: Manolis Surligas <surligas@gmail.com>